fix an issue with setup.py

+++ utils/make_release.py
@@ -102,6 +102,7 @@ if __name__ == "__main__":
             
         if not os.path.exists(setup_path):
             # no setup.py in new grid2op
+            does_setup_exists = False
             with open(grid2op_init, "r") as f:
                 new_setup = f.read()
             try:
@@ -112,6 +113,7 @@ if __name__ == "__main__":
             #     "script \"update_version\" cannot find the root path of Grid2op. "
             #     "Please provide a valid \"--path\" argument.")
         else:
+            does_setup_exists = True
             with open(setup_path, "r") as f:
                 new_setup = f.read()
             try:
@@ -134,11 +136,13 @@ if __name__ == "__main__":
                            "version='{}'".format(version),
                            new_setup)
         
-        with open(setup_path, "w") as f:
-            f.write(new_setup)
+        if does_setup_exists:
+            # setup.py exists
+            with open(setup_path, "w") as f:
+                f.write(new_setup)
 
-        # Stage in git
-        start_subprocess_print(["git", "add", setup_path])
+            # Stage in git
+            start_subprocess_print(["git", "add", setup_path])
 
         # grid2op/__init__.py
         with open(grid2op_init, "r") as f:
